requirements.txt
django_smtp_ssl
django-mailgun
django-anymail[mailgun]